# backend/swagger/api-docs/ram-api-docs.yaml
swagger: '2.0'
...
info:
  title: RAM Public API
  description: Full documentation of Relationship Access Manager (RAM) public API. All the listed API calls must be prefixed with /api/v1
  version: "1.0.0"
host:
schemes:
  - http
  - https
basePath: /api/v1
produces:
  - application/json

# ---------------------------------------------------------------------------------------------------------------------
# End Points
# ---------------------------------------------------------------------------------------------------------------------
paths:
...
  /party/identity/{idValue}:
    get:
      tags:
        - Party
      operationId: Party.findByIdentityIdValue
      summary: Get the party by its identity id value
      description: Get the party by its identity id value
      parameters:
        - $ref: "#/parameters/idValue"
      responses:
        200:
          description: The updated party returned
          schema:
            $ref: "#/definitions/Party"
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'

  # Party Types .......................................................................................................
...
  /relationship/{id}:
    get:
      tags:
        - Relationship
      operationId: Relationship.findByIdentifier
      summary: Get Relationship
      description: Get Relationship
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        200:
          schema:
            $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'
    put:
      tags:
        - Relationship
      operationId: Relationship.modify
      summary: Update Relationship
      description: Update Relationship
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - in: body
          required: true
          schema:
            $ref: '#/definitions/Relationship'
      responses:
        200:
          schema:
            $ref: '#/definitions/Relationship'
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'
...
  /relationships/identity/{idValue}:
    get:
      tags:
        - Relationship
      operationId: Relationship.searchByIdentity
      summary: Search Relationships for an identity
      description: Search Relationships for an identity
      parameters:
        - $ref: "#/parameters/idValue"
        - name: filter
          in: query
          description: Filtering and Sorting
          required: false
          type: string
        - name: page
          in: query
          description: Page used for pagination
          required: true
          type: number
          format: int32
        - name: pageSize
          in: query
          description: Page Size used for pagination
          required: false
          type: number
          format: int32
      responses:
        200:
          schema:
            $ref: '#/definitions/RelationshipListResponse'
        400:
          $ref: '#/responses/400'
        500:
          $ref: '#/responses/500'
...
  /relationshipStatuses:
    get:
      tags:
        - Reference - Relationship Status
      operationId: Relationship.listStatuses
      summary: Get the relationship statuses
      description: Get the relationship statuses
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/RelationshipStatusResourceReference'
        500:
          $ref: '#/responses/500'

  # -------------------------------------------------------------------------------------------------------------------
  # Reference End Points
  # -------------------------------------------------------------------------------------------------------------------

  # Relationship Types ................................................................................................
...
  /role:
    post:
      tags:
        - Role
      operationId: Role.createByIdentity
      summary: Create a role with an agency service
      description: Create a role with an agency service as an agency user
      parameters:
        - in: body
          required: true
          schema:
            $ref: '#/definitions/Role'
      responses:
        200:
          schema:
            $ref: '#/definitions/Role'
        400:
          $ref: '#/responses/400'
        404:
          $ref: '#/responses/404'
        500:
          $ref: '#/responses/500'


  # Role Types ........................................................................................................
...
definitions:
...
  Profile:
    type: object
    description: |
      A profile is a collection of detail about the party. A myGov profile consists of Name and Date of Birth. The
      CSP may or may not supply the profile.  Not all identities will be supplied by the CSP. Some profiles may
      be self asserted.
    required: ["provider","name"]
    properties:
      provider:
        type: string
        description: The provider of the profile
        enum: ['ABR','AUTHENTICATOR_APP','MY_GOV','SELF_ASSERTED','VANGUARD','TEMP']
      name:
        $ref: '#/definitions/Name'
      sharedSecrets:
        type: array
        items:
          $ref: '#/definitions/SharedSecret'

  # Name ..............................................................................................................
  Name:
    type: object
    description: Name of a party of profile for either an individual or organisation
    properties:
      givenName:
        type: string
        description: The given name of an individual
      familyName:
        type: string
        description: The family name of an individual
      unstructuredName:
        type: string
        description: The unstructured name of an organisation
      _displayName:
        type: string
        description: The computed name of the individual or organisation

  # Shared Secret .....................................................................................................
  SharedSecret:
    type: object
    description: A Shared Secret is a record of sensitive information pertaining to an identity
    properties:
      value:
        type: string
        description: Hashed Value
      sharedSecretType:
        $ref: '#/definitions/SharedSecretType'
  SharedSecretType:
    allOf:
      - $ref: '#/definitions/ICodeDecode'
    type: object
    description: The type of a Shared Secret
    properties:
      domain:
        type: string
        description: The domain the shared secrety type is for

  # Party .............................................................................................................
  Party:
    allOf:
    - $ref: '#/definitions/IRAMObject'
    - $ref: '#/definitions/PartyDetails'
    description: A Party is the concept that participates in Relationships.
    type: object
    required: ["roles","identities"]
  PartyDetails:
    type: object
    required: ["partyType","identities"]
    properties:
      partyType:
        type: string
      identities:
        type: array
        items:
          $ref: '#/definitions/IdentityResourceReference'
  PartyResourceReference:
    description: A reference to a Party
    type: object
    required: ["href"]
    properties:
      href:
        type: string
        description: URI that uniquely identifies this Party
      value:
          $ref: '#/definitions/Party'
...
  Identity:
    allOf:
      - $ref: '#/definitions/IRAMObject'
      - $ref: '#/definitions/IdentityBusinessIdentifier'
      - $ref: '#/definitions/IdentityDetails'
    type: object
    description: |
      Parties have at least one, and possibly many identities.  There will be just one default identity.
      Identities consist of a profile which conists of a Name & maybe recorded "shared secret".
      It is the name from the default identity that will be used when it is not possible  to use the
      authentication details to refer to a Party, e.g. when displaying a list of related parties.
    required: ["idType","idValue","defaultIndictor", "profile"]
  IdentityBusinessIdentifier:
    type: object
    required: ["idValue","rawIdValue","identityType"]
    properties:
      idValue:
        description: The identity value (as a composite of the type, scheme, raw id value)
        type: string
      rawIdValue:
        description: The raw identity value
        type: string
      identityType:
        type: string
        description: The identity type
        enum: ['AGENCY_PROVIDED_TOKEN','INVITATION_CODE','LINK_ID','PUBLIC_IDENTIFIER']
  IdentityDetails:
    type: object
    required: ["defaultInd","profile","party"]
    properties:
      defaultInd:
        type: boolean
        description: The default indicator
      strength:
        type: number
        description: The strength
      agencyScheme:
        type: string
        description: The agency scheme
      agencyToken:
        type: string
        description: The agency token
      invitationCodeStatus:
        type: string
        description: The invitation code status
        enum: ['CLAIMED','PENDING','REJECTED']
      invitationCodeExpiryTimestamp:
        type: string
        description: The invitation code expiry timestamp
      invitationCodeClaimedTimestamp:
        type: string
        description: The invitation code claimed timestamp
      invitationCodeTemporaryEmailAddress:
        type: string
        description: The invitation code temporary email address
      publicIdentifierScheme:
        type: string
        description: The public identifier scheme
      linkIdScheme:
        type: string
        description: The link id scheme
      linkIdConsumer:
        type: string
        description: The link id consumer
      profile:
        $ref: '#/definitions/Profile'
      party:
        # todo resource reference or direct party?
        $ref: '#/definitions/Party'
  IdentityResourceReference:
    type: object
    description: A reference to a Resource
    required: ["href"]
    properties:
      href:
        type: string
        description: URI that uniquely identifies this Identity
      value:
        $ref: '#/definitions/Identity'

  # Relationship ......................................................................................................
  Relationship:
    allOf:
    - $ref: '#/definitions/IRAMObject'
    - $ref: '#/definitions/RelationshipBusinessIdentifier'
    - $ref: '#/definitions/RelationshipDetails'
    type: object
    description: |
      A Relationship is a connection between two parties.
      RAM will only record relationships if it is useful (in some way) to agencies for deciding if one party to
      confer authoriatation to another party
    required: ["relationshipTypeURI","subjectPartyRef","delegatePartyRef","startTimestamp","status"]
  RelationshipBusinessIdentifier:
    type: object
    required: ["relationshipTypeURI","subjectPartyRef","delegatePartyRef","startTimestamp"]
    properties:
      relationshipTypeURI:
        type: string
        description: URI that uniquely identifies the reference data type that gives semantic meaning to  this relationship
      subjectPartyRef:
        $ref: '#/definitions/PartyResourceReference'
      delegatePartyRef:
        $ref: '#/definitions/PartyResourceReference'
      startTimestamp:
        type: string
        description: Defines when this relationship comes into effect.  Date in ISO format.
        format: date
  RelationshipDetails:
    type: object
    required: ["status"]
    properties:
      code:
        type: string
        description: Unique code
      # this has not yet been standardised to a role reference
      subjectPartyRoleURI:
        type: string
        description: URI that uniquely identifies the subject's role in this relationship
      delegatePartyRoleURI:
      # this has not yet been standardised to a role reference
        type: string
        description: URI that uniquely identifies the delegate's role in this relationship
      subjectsNickname:
          $ref: '#/definitions/Name'
      delegatesNickname:
          $ref: '#/definitions/Name'
      endTimestamp:
        type: string
        description: Date in ISO format
        format: date
      attributes:
        type: array
        description: Attributes attached to this relationship based on role type
        items:
          $ref: '#/definitions/RelationshipAttribute'
      status:
        type: string
        description: Indicates where in the lifecycle of relationships; this one is up to.
        enum: ['invalid','pending','active','deleted','cancelled']
...
  RelationshipResourceReference:
    type: object
    description: A reference to a Resource
    required: ["href"]
    properties:
      href:
        type: string
        description: URI that uniquely identifies this Relationship
      value:
        $ref: '#/definitions/Relationship'
  RelationshipListResponse:
    allOf:
    - $ref: '#/definitions/SearchResultResponse'
    type: object
    required: ["list"]
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/RelationshipResourceReference'

  # Relationship Attribute ............................................................................................
  RelationshipStatus:
    type: object
    description: Relationship status
    properties:
      code:
        type: string
        description: The code or name
      shortDecodeText:
        type: string
        description: The human friendly label
  RelationshipStatusResourceReference:
    type: object
    description: A reference to a Resource
    required: ["href"]
    properties:
      href:
        type: string
        description: URI that uniquely identifies this RelationshipStatus
      value:
        $ref: '#/definitions/RelationshipStatus'

  # Relationship Attribute ............................................................................................
  RelationshipAttribute:
    allOf:
    - $ref: '#/definitions/IRAMObject'
    type: object
    description: A Relationship Attribute is a way of recording some "facts" about a Relationship
    required: ["name","value"]
    properties:
      name:
        type: string
        description: Name of an attribute definition
      attributeName:
        $ref: '#/definitions/RelationshipAttributeNameResourceReference'

  # -------------------------------------------------------------------------------------------------------------------
  # Reference Models
  # -------------------------------------------------------------------------------------------------------------------

  # Relationship Type .................................................................................................
...
  RelationshipAttributeName:
    allOf:
      - $ref: '#/definitions/ICodeDecode'
      - $ref: '#/definitions/RelationshipAttributeNameBusinessIdentifier'
      - $ref: '#/definitions/RelationshipAttributeNameDetails'
  RelationshipAttributeNameBusinessIdentifier:
    type: object
    required: ["code"]
    properties:
      name:
        type: string
        description: This is the short code value by which Relationship Attribute Names may be known
  RelationshipAttributeNameDetails:
    type: object
    required: ["optionalInd","fieldType"]
    properties:
      domain:
        description: It is the field type
        type: string
        enum: ['NULL','BOOLEAN','NUMBER','STRING','DATE','MARKDOWN','SINGLE_SELECT','MULTI_SELECT']
      classifier:
        description: It is the classifier to indicate what type of attribute this is and how it is meant to be used (eg this is a permission)
        type: string
        enum: ['PERMISSION', 'OTHER']
      category:
        description: An optional category or grouping for additional meta data (eg a permission in Administrative Services)
        type: string
      purposeText:
        description: An optional purpose text to indicate how this attribute is meant to be used
        type: string
      permittedValues:
        description: An array of the accepted values
        type: array
        items:
          type: string
  RelationshipAttributeNameResourceReference:
    type: object
    description: A reference to a RelationshipAttributeName Resource
    required: ["href"]
    properties:
      href:
        type: string
        description: URL that uniquely identifies this RelationshipAttributeName
      value:
        description: The content of this field is optional, it is the same content for a get request on the href
        schema:
          $ref: '#/definitions/RelationshipAttributeName'

  # Role ..............................................................................................................
  Role:
    allOf:
    - $ref: '#/definitions/IRAMObject'
    type: object
    description: Role
    properties:
      code:
        type: string
        description: Unique code
      roleType:
        $ref: '#/definitions/RoleTypesResourceReference'
      party:
        $ref: '#/definitions/PartyResourceReference'
      attributes:
        type: array
        description: Role attributes that may be used with this resource
        items:
          $ref: '#/definitions/RoleAttribute'
  RoleAttribute:
    type: object
    description: Role Attribute
    required: ["value", "attributeName"]
    properties:
      value:
        type: string
        description: Value of Role Attribute
      attributeName:
        description: Role Attribute Name Resource Reference
        schema:
          $ref: '#/definitions/RoleAttributeNameResourceReference'
...
  RoleType:
    allOf:
      - $ref: '#/definitions/ICodeDecode'
      - $ref: '#/definitions/RoleTypeAttributeNameBusinessIdentifier'
      - $ref: '#/definitions/RoleTypeAttributeNameDetails'
  RoleTypeAttributeNameBusinessIdentifier:
    type: object
    required: ["code"]
    properties:
      code:
        type: string
        description: This is the short code value by which Relationship Types may be known
  RoleTypeAttributeNameDetails:
    type: object
    required: ["roleAttributeNames"]
    properties:
      roleAttributeNames:
        type: array
        description: Role attributes that may be used with this resource
        items:
          $ref: '#/definitions/RoleAttributeNameUsage'
  RoleTypesResourceReference:
    type: object
    description: A reference to a RoleType Resource
    required: ["href"]
    properties:
      href:
        type: string
        description: URL that uniquely identifies this RoleType
      value:
        description: The content of this field is optional, it is the same content for a get request on the href
        schema:
          $ref: '#/definitions/RoleType'

  # Role Attribute Name Usage .........................................................................................
  RoleAttributeNameUsage:
    allOf:
      - $ref: '#/definitions/RoleAttributeNameUsageDetails'
  RoleAttributeNameUsageDetails:
    type: object
    required: ["optionalInd","defaultValue"]
    properties:
      optionalInd:
        description: Indicator of whether the attribute is optional or not
        type: boolean
      defaultValue:
        description: The default value of the attribute
        type: string
      attributeNameDef:
        description: The attribute name
        $ref: '#/definitions/RoleAttributeNameResourceReference'

  # Role Attribute Name ...............................................................................................
  RoleAttributeName:
    allOf:
      - $ref: '#/definitions/ICodeDecode'
      - $ref: '#/definitions/RoleAttributeNameBusinessIdentifier'
      - $ref: '#/definitions/RoleAttributeNameDetails'
  RoleAttributeNameBusinessIdentifier:
    type: object
    required: ["code"]
    properties:
      name:
        type: string
        description: This is the short code value by which Role Attribute Names may be known
  RoleAttributeNameDetails:
    type: object
    required: ["optionalInd","fieldType"]
    properties:
      domain:
        description: It is the field type
        type: string
        enum: ['NULL','BOOLEAN','NUMBER','STRING','DATE','MARKDOWN','SINGLE_SELECT','MULTI_SELECT']
      classifier:
        description: It is the classifier to indicate what type of attribute this is and how it is meant to be used (eg this is a permission)
        type: string
        enum: ['AGENCY_SERVICE', 'PERMISSION', 'OTHER']
      category:
        description: An optional category or grouping for additional meta data (eg a permission in Administrative Services)
        type: string
      purposeText:
        description: An optional purpose text to indicate how this attribute is meant to be used
        type: string
      permittedValues:
        description: An array of the accepted values
        type: array
        items:
          type: string
  RoleAttributeNameResourceReference:
    type: object
    description: A reference to a RoleAttributeName Resource
    required: ["href"]
    properties:
      href:
        type: string
        description: URL that uniquely identifies this RoleAttributeName
      value:
        description: The content of this field is optional, it is the same content for a get request on the href
        schema:
          $ref: '#/definitions/RoleAttributeName'
...
  IRAMObject:
    type: object
    required: ['createdAt','updatedAt','deleteInd','resourceVersion']
    properties:
      createdAt:
        description: A timestamp when this resource was first created.
        type: string
        format: date-time
      updatedAt:
        description: A timestamp when this resource was last updated.
        type: string
        format: date-time
      deleteInd:
        type: boolean
        description: An indicated if this resource has been logically deleted.
      resourceVersion:
        type: number
        description: A schema version.
  ICodeDecode:
    type: object
    required: ['code','shortDecodeText','longDecodeText','startTimestamp']
    properties:
      code:
        description: A human friendly mneumonic for this reference data concept
        type: string
      shortDecodeText:
        description: This is a short label for this item of reference data
        type: string
      longDecodeText:
        type: string
        description:  This is a long label for this item of reference data
      startTimestamp:
        type: string
        description: This is the timestamp at which this reference data item may start being used
        format: date-time
      endTimestamp:
        type: string
        description: This is the timestamp at which this reference data item ceased being able to be used (but may continue to exist on existing records)
        format: date-time
  SearchResultResponse:
    type: object
    description: |
      This abstract object just adds common properties to a <entity>SearchResultResponse object
    required: [totalCount, pageSize]
    properties:
      page:
        description: Current page number
        type: integer
        format: int32
      totalCount:
        description: Total number of records matched for this query
        type: integer
        format: int32
      pageSize:
        description: Page Size for the results, which the server may ignore for performance reasons
        type: integer
        format: int32
  Error:
    type: object
    required: ["alert"]
    properties:
      alert:
        $ref: '#/definitions/Alert'
  Alert:
    type: object
    required: ["alertType","messages"]
    properties:
        alertType:
            type: string
            format: can be one of the following values, success, notice or error
        messages:
            type: array
            description: Array of string presenting one or more messages
            items:
                type: string

# ---------------------------------------------------------------------------------------------------------------------
# Parameters
# ---------------------------------------------------------------------------------------------------------------------
parameters:
...
  idValue:
    name: idValue
    in: path
    description: The unique identity composite id value
    required: true
    type: string
...
responses:
  400:
    description: Bad Input Error
    schema:
      $ref: '#/definitions/Error'
...
  404:
    description: Not Found Error
    schema:
      $ref: '#/definitions/Error'
  500:
    description: Server Error
    schema:
      $ref: '#/definitions/Error'